[FIX] web_editor: fix error when saving button changed to link
Problem:
Saving the button snippet after changing its type to a link throws an
error because `snippet_key` is `undefined`.

Cause:
The button snippet editor should be disabled if the button is changed
to a link. However, `updateOptionsUIVisibility` is only triggered on
click events. In this case, the element type changes through the
editor itself, not by clicking, so the UI isn't updated accordingly.

Solution:
Trigger a `click` event on the link element programmatically to call
`updateOptionsUIVisibility` and hide the button snippet editor when
the element is no longer a button.

Steps to reproduce:
- Add a button snippet
- Click inside the button to edit
- Notice the Button snippet editor appears
- Change the type to "Link" instead of "Primary"
- The Button snippet editor is still visible
- Click the floppy disk icon (save) in the snippet editor
-> A traceback occurs

opw-4936796

closes odoo/odoo#220787

X-original-commit: 6f76133fbfdec4a26755a4f2e428b289cd60e897
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>
Signed-off-by: Walid Sahli (wasa) <wasa@odoo.com>