[FIX] account,account_edi_ubl_cii: Fix get invoice legal document
`_get_invoice_legal_documents` should, and is expected to, return a dict. however, if called with `filetype = all`, it returns, because of `_get_invoice_legal_documents_all`, a list which breaks calling code as they expect a dict not a list, and this part of the code is not used anywhere nor tested.

- remove the line causing `_get_invoice_legal_documents` to return a list.

- make `download_invoice_documents_filetype` work with multiple filetypes

no-task

closes odoo/odoo#220170

Signed-off-by: John Laterre (jol) <jol@odoo.com>